--- routes/articleCRUD/createArticle.js
@@ -1,19 +1,0 @@
-import express from "express";
-import { PrismaClient } from "@prisma/client";
-import asyncHandler from "../asyncHandler.js";
-
-const router = express.Router();
-const prisma = new PrismaClient();
-
-router.post(
-  "/",
-  asyncHandler(async (req, res) => {
-    const { title, content } = req.body;
-    const article = await prisma.article.create({
-      data: { title, content },
-    });
-    res.status(201).json(article);
-  })
-);
-
-export default router;
