--- routes/articleCRUD/updateArticle.js
@@ -1,24 +1,0 @@
-import express from "express";
-import { PrismaClient } from "@prisma/client";
-import asyncHandler from "../asyncHandler.js";
-
-const router = express.Router();
-const prisma = new PrismaClient();
-
-router.patch(
-  "/:id",
-  asyncHandler(async (req, res) => {
-    const { id } = req.params;
-    const { title, content } = req.body;
-    const article = await prisma.article.update({
-      where: { id: parseInt(id) },
-      data: { title, content },
-    });
-    if (!article) {
-      return res.status(404).json({ error: "Article not found" });
-    }
-    res.status(200).json(article);
-  })
-);
-
-export default router;
