--- routes/productCRUD/updateProduct.js
@@ -1,24 +1,0 @@
-import express from "express";
-import { PrismaClient } from "@prisma/client";
-import asyncHandler from "../asyncHandler.js";
-
-const router = express.Router();
-const prisma = new PrismaClient();
-
-router.patch(
-  "/:id",
-  asyncHandler(async (req, res) => {
-    const { id } = req.params;
-    const { name, description, price, tags } = req.body;
-    const product = await prisma.product.update({
-      where: { id: parseInt(id) },
-      data: { name, description, price, tags },
-    });
-    if (!product) {
-      return res.status(404).json({ error: "Product not found" });
-    }
-    res.json(product);
-  })
-);
-
-export default router;
