--- routes/productCommentsCRUD/deleteProductComment.js
@@ -1,19 +1,0 @@
-import express from "express";
-import { PrismaClient } from "@prisma/client";
-import asyncHandler from "../asyncHandler.js";
-
-const prisma = new PrismaClient();
-const router = express.Router();
-
-router.delete(
-  "/:id",
-  asyncHandler(async (req, res) => {
-    const { id } = req.params;
-    await prisma.productComment.delete({
-      where: { id: parseInt(id) },
-    });
-    res.status(204).end();
-  })
-);
-
-export default router;
