+++ service/productCommentService.js
@@ -1,0 +1,65 @@
+import express from "express";
+import { PrismaClient } from "@prisma/client";
+import asyncHandler from "../Common/asyncHandler.js";
+
+const prisma = new PrismaClient();
+const router = express.Router();
+
+router.post(
+  "/:productId",
+  asyncHandler(async (req, res) => {
+    const { productId } = req.params;
+    const { content } = req.body;
+
+    const comment = await prisma.productComment.create({
+      data: {
+        content,
+        productId: parseInt(productId),
+      },
+    });
+    res.status(201).json(comment);
+  })
+);
+
+router.get(
+  "/:productId",
+  asyncHandler(async (req, res) => {
+    const { productId } = req.params;
+    const { cursor } = req.query;
+
+    const comments = await prisma.productComment.findMany({
+      where: { productId: parseInt(productId) },
+      take: 10,
+      skip: cursor ? 1 : 0,
+      cursor: cursor ? { id: parseInt(cursor) } : undefined,
+    });
+
+    res.status(200).json(comments);
+  })
+);
+
+router.patch(
+  "/:id",
+  asyncHandler(async (req, res) => {
+    const { id } = req.params;
+    const { content } = req.body;
+    const comment = await prisma.productComment.update({
+      where: { id: parseInt(id) },
+      data: { content },
+    });
+    res.status(200).json(comment);
+  })
+);
+
+router.delete(
+  "/:id",
+  asyncHandler(async (req, res) => {
+    const { id } = req.params;
+    await prisma.productComment.delete({
+      where: { id: parseInt(id) },
+    });
+    res.status(204).end();
+  })
+);
+
+export default router;
